requirements.txt
tensorflow
tensorflow-serving-api
absl-py
